renamed authentication to middleware and changed post router
