+++ src/middleware/authMiddleware.js
@@ -1,33 +1,35 @@
 const jwt = require("jsonwebtoken");
-const { decodeJWT, findToken } = require('../utilities/jwtUtilities');
 
-async function authenticateToken(req, res, next){
-    const token = findToken(req);
-    if (!token){
-        res.status(401).json({message: "You must sign in first"});
-        return;
+
+const authenticate = (req, res, next) => {
+    const token = req.headers?.authorization && req.headers.authorization.split(" ")[1];
+    if (!token) {
+        return res.status(401).json("Unauthorized Access");
     }
-    const user = decodeJWT(token);
-    req.User = user;
-    next();
+    try {
+        const user = jwt.verify(token, process.env.JWT_SECRET);
+        res.locals.user = user;
+        next();
+    } catch (err) {
+        return res.status(401).json("Unauthorized Access, try relogging");
+    }
 }
 
-async function authenticateAdminToken(req, res, next){
-    const token = findToken(req);
-    if (!token){
-        res.status(401).json({message: "You must sign in first"});
-        return;
+const adminAuthenticate = (req, res, next) => {
+    const token = req.headers?.authorization && req.headers.authorization.split(" ")[1];
+    if (!token) {
+        return res.status(401).json("Unauthorized Access");
     }
-    const user = decodeJWT(token);
-    if (!user || user.Role !== "Admin"){
-        res.status(403).json({message: "Admin only access"});
-        return;
+    try {
+        const user = jwt.verify(token, process.env.JWT_SECRET);
+        if (user.role !== "admin") {
+            return res.status(401).json({message: "Privilege too low"});
+        }
+        res.locals.user = user;
+        next();
+    } catch (err) {
+        return res.status(401).json("Unauthorized Access, try relogging");
     }
-    req.User = user;
-    next();
 }
 
-module.exports = {
-    authenticateToken,
-    authenticateAdminToken
-};
+module.exports = {authenticate, adminAuthenticate};
